feat: Use the navigation service for the import view model
