+++ WpfGui/Interfaces/INavigationService.cs
@@ -12,8 +12,10 @@ public interface INavigationService
     AddUrlDialogResult ShowAddUrlView(MainViewModel mainViewModel);
     void ShowBalanceView(ViewModel api);
     ItemDialogResult ShowEditView(MainViewModel mainViewModel);
+    FileDialogResult ShowImportFileDialog(ImportFileType fileType);
     void ShowImportView(ViewModel api);
     bool? ShowMessageBox(string message, string title, bool allowCancel = false);
     FileDialogResult ShowOpenFileDialog(bool checkFileExists = true);
+    FolderDialogResult ShowOpenFolderDialog(string folder);
     void ShowRankView(List<ActionViewModel> items, ViewModel api);
 }
